+++ json-path/src/test/java/com/jayway/jsonpath/old/FilterTest.java
@@ -191,7 +191,6 @@ public class FilterTest extends BaseTest {
     }
 
     @Test
-    //@Ignore //TODO: finalize behaviour
     public void exists_filters_evaluates() throws Exception {
         Map<String, Object> check = new HashMap<String, Object>();
         check.put("foo", "foo");
@@ -200,8 +199,8 @@ public class FilterTest extends BaseTest {
         assertTrue(filter(where("foo").exists(true)).apply(createPredicateContext(check)));
         assertFalse(filter(where("foo").exists(false)).apply(createPredicateContext(check)));
 
-        assertTrue(filter(where("foo_null").exists(true)).apply(createPredicateContext(check)));
-        assertFalse(filter(where("foo_null").exists(false)).apply(createPredicateContext(check)));
+        assertTrue(filter(where("foo_null").exists(false)).apply(createPredicateContext(check)));
+        assertFalse(filter(where("foo_null").exists(true)).apply(createPredicateContext(check)));
 
         assertTrue(filter(where("bar").exists(false)).apply(createPredicateContext(check)));
         assertFalse(filter(where("bar").exists(true)).apply(createPredicateContext(check)));